moving strategy interface
