+++ src/bloqade/lanes/analysis/placement/__init__.py
@@ -1,11 +1,13 @@
 from .analysis import (
     PlacementAnalysis as PlacementAnalysis,
-    PlacementStrategyABC as PlacementStrategyABC,
-    SingleZonePlacementStrategyABC as SingleZonePlacementStrategyABC,
 )
 from .lattice import (
     AtomState as AtomState,
     ConcreteState as ConcreteState,
     ExecuteCZ as ExecuteCZ,
     ExecuteMeasure as ExecuteMeasure,
+)
+from .strategy import (
+    PlacementStrategyABC as PlacementStrategyABC,
+    SingleZonePlacementStrategyABC as SingleZonePlacementStrategyABC,
 )
