+++ src/bloqade/lanes/analysis/placement/strategy.py
@@ -1,0 +1,103 @@
+import abc
+
+from bloqade.lanes.layout import LaneAddress, LocationAddress, ZoneAddress
+
+from .lattice import AtomState, ConcreteState, ExecuteCZ, ExecuteMeasure
+
+
+class PlacementStrategyABC(abc.ABC):
+
+    @abc.abstractmethod
+    def validate_initial_layout(
+        self,
+        initial_layout: tuple[LocationAddress, ...],
+    ) -> None: ...
+
+    @abc.abstractmethod
+    def cz_placements(
+        self,
+        state: AtomState,
+        controls: tuple[int, ...],
+        targets: tuple[int, ...],
+    ) -> AtomState: ...
+
+    @abc.abstractmethod
+    def sq_placements(
+        self,
+        state: AtomState,
+        qubits: tuple[int, ...],
+    ) -> AtomState: ...
+
+    @abc.abstractmethod
+    def measure_placements(
+        self,
+        state: AtomState,
+        qubits: tuple[int, ...],
+    ) -> AtomState: ...
+
+
+class SingleZonePlacementStrategyABC(PlacementStrategyABC):
+
+    @abc.abstractmethod
+    def compute_moves(
+        self,
+        state_before: ConcreteState,
+        state_after: ConcreteState,
+    ) -> tuple[tuple[LaneAddress, ...], ...]: ...
+
+    @abc.abstractmethod
+    def desired_cz_layout(
+        self,
+        state: ConcreteState,
+        controls: tuple[int, ...],
+        targets: tuple[int, ...],
+    ) -> ConcreteState: ...
+
+    def cz_placements(
+        self, state: AtomState, controls: tuple[int, ...], targets: tuple[int, ...]
+    ) -> AtomState:
+        if len(controls) != len(targets) or state == AtomState.bottom():
+            return AtomState.bottom()
+
+        if not isinstance(state, ConcreteState):
+            return AtomState.top()
+
+        desired_state = self.desired_cz_layout(state, controls, targets)
+        move_layers = self.compute_moves(state, desired_state)
+
+        return ExecuteCZ(
+            occupied=state.occupied,
+            layout=desired_state.layout,
+            move_count=tuple(
+                mc + int(src != dst)
+                for mc, src, dst in zip(
+                    state.move_count, state.layout, desired_state.layout
+                )
+            ),
+            active_cz_zones=frozenset(
+                [ZoneAddress(0)]
+            ),  # Assuming single zone with address 0
+            move_layers=move_layers,
+        )
+
+    def sq_placements(self, state: AtomState, qubits: tuple[int, ...]) -> AtomState:
+        return state  # No movement needed for single zone
+
+    def measure_placements(
+        self, state: AtomState, qubits: tuple[int, ...]
+    ) -> AtomState:
+        if not isinstance(state, ConcreteState):
+            return state
+
+        # all qubits must be measured
+        if len(qubits) != len(state.layout):
+            return AtomState.bottom()
+
+        return ExecuteMeasure(
+            occupied=state.occupied,
+            layout=state.layout,
+            move_count=state.move_count,
+            zone_maps=tuple(
+                ZoneAddress(0) for _ in qubits
+            ),  # Assuming single zone with address 0
+        )
